refactor(menu): split message_reader.go into concern-grouped files
message_reader.go was 1,684 lines. Extract its 21 helper functions into
four sibling files (same package, pure code movement, no behavior change):

- message_reader_format.go   - line/body classification + ANSI-art detection
- message_reader_subst.go    - template token substitution + auto-widths
- message_reader_nav.go      - reader navigation actions (reply/thread/jump/help)
- message_reader_headers.go  - MSGID/header discovery + header-type prompt

message_reader.go drops to ~690 lines (runMessageReader and its loop).
Imports are grouped stdlib/external per file. go vet clean, go build ./...
clean, 269 menu tests pass (race-clean), 1423 total across 49 packages.
